[stable/spinnaker] fix bug where spinnaker can't connect to minio (#14866)
* [stable/spinnaker] fix bug where spinnaker can't connect to minio

As reported in spinnaker itself [here](https://github.com/spinnaker/spinnaker/issues/4431),
It doesn't appear to affect the current stable build, but it has problems with the nightly builds.

This change should innoculate us from that bug if it makes it through to a stable build.

Effectively the issue is that in the newer builds of spinnaker it attempts to use domain paths
for s3 access.  This means instead of trying to connect to `spinnaker-minio/spinnaker` which should
be valid dns for the minio service it tries to connect to `spinnaker.spinnaker-minio` and fails because
this is not valid dns for minio.

Signed-off-by: Paul Czarkowski <username.taken@gmail.com>

* bump halyard version for cli arg needed

Signed-off-by: Paul Czarkowski <username.taken@gmail.com>

* fix argument

Signed-off-by: Paul Czarkowski <username.taken@gmail.com>

* fix version

Signed-off-by: Paul Czarkowski <username.taken@gmail.com>